finalizing tests

+++ controllers/controller.js
@@ -24,7 +24,12 @@ class Controller {
   //!!! KERJAINNNNN============
   static async profilePage(req, res) {
     try {
+      // if (!req.session.user) {
+      //   const error = "Please log in to view your profile!";
+      //   return res.redirect(`/login?error=${error}`);
+      // }
       // res.send(`profile page!!`);
+      console.log(req.session.user);
       const { id } = req.params;
       const user = await User.findByPk(id, {
         include: { model: UserProfile },
@@ -41,13 +46,29 @@ class Controller {
   //! UDAH
   static async companiesPage(req, res) {
     try {
-      const { search, category } = req.query;
+      //!!CEK SI USER PUNYA INVESTMENT ITU GAAA
+      let userInvestments = [];
+      if (req.session.user) {
+        const userId = req.session.user.id;
+        userInvestments = await Investment.findAll({
+          where: { UserId: userId },
+          attributes: ["id", "CompanyId"],
+        });
+      }
+      const investmentMap = userInvestments.reduce((acc, inv) => {
+        acc[inv.CompanyId] = inv.id;
+        return acc;
+      }, {});
+      // console.log(userInvestments, `INI NIHH`, investedCompanyIds);
+      //UDAHHH
+
+      const { search, category, error } = req.query;
       let filter = {};
       let categoryFilter = {};
 
       if (search) {
         filter.name = {
-          [Op.iLike]: `%${search}%`, // Case-insensitive search
+          [Op.iLike]: `%${search}%`,
         };
       }
 
@@ -67,7 +88,15 @@ class Controller {
       const categories = await Category.findAll();
 
       // console.log(categories);
-      res.render("companiesPage", { companies, categories, search, category });
+      res.render("companiesPage", {
+        companies,
+        categories,
+        search,
+        category,
+        //TAMBAHIN YG DIPUNYAA
+        investmentMap,
+        error: error || null,
+      });
     } catch (err) {
       console.log(err);
       res.send(err);
@@ -114,7 +143,6 @@ class Controller {
       const { id } = req.params;
       const { name, description, UserId, investmentType, amount } = req.body;
 
-
       // console.log(req.body);
 
       const newInvestment = await Investment.create({
@@ -136,7 +164,17 @@ class Controller {
 
   static async userInvestments(req, res) {
     try {
+      const { id, role } = req.session.user;
+
+      //!!!!!!!KONDISI BUAT ADMIN
+      let condition = {};
+      if (role !== "admin") {
+        condition = { id };
+      }
+      //! UDAH
+
       const investments = await User.findAll({
+        where: condition,
         include: {
           model: Company,
           through: {
@@ -150,7 +188,6 @@ class Controller {
           },
         },
       });
-
       // const investments = await Investment.findAll({
       //   include: {
       //     model: User,
@@ -159,7 +196,6 @@ class Controller {
       //     }
       //   }
       // })
-
       // res.send(investments);
       res.render("investments", { investments, formatRupiah, formatDate });
     } catch (err) {
@@ -170,6 +206,8 @@ class Controller {
 
   static async editInvestmentForm(req, res) {
     const { investmentId } = req.params;
+    const { id: loggedInUserId, role } = req.session.user;
+
     try {
       const companies = await Company.findAll();
       const investment = await Investment.findOne({
@@ -181,8 +219,18 @@ class Controller {
           "investmentType",
           "amount",
           "CompanyId",
+          "UserId",
         ],
       });
+
+      console.log(investment.UserId, `bandingin sama `, loggedInUserId);
+      console.log(investment);
+      //CHECK ROLEE SAMA USER
+      if (role !== "admin" && investment.UserId !== loggedInUserId) {
+        let error = `no access!`;
+        return res.redirect(`/companies?error=${error}`);
+      }
+      //
 
       res.render("editInvestment", { investment, companies });
       // res.send(companies);
@@ -195,12 +243,21 @@ class Controller {
 
   static async updateInvestment(req, res) {
     const { investmentId } = req.params;
+    const { id: loggedInUserId, role } = req.session.user;
+
     const { name, description, investmentType, amount, CompanyId } = req.body;
     try {
-      await Investment.update(
+      const investment = await Investment.update(
         { name, description, investmentType, amount, CompanyId },
         { where: { id: investmentId } }
       );
+
+      // CHECK ROLEE SAMA USER
+      if (role !== "admin" && investment.UserId !== loggedInUserId) {
+        let error = `no access!`;
+        return res.redirect(`/companies?error=${error}`);
+      }
+      //
 
       // res.send(`UPDATE BERHASIL!`);
       res.redirect("/investments");
@@ -212,7 +269,19 @@ class Controller {
 
   static async deleteInvestment(req, res) {
     const { investmentId } = req.params;
+    const { id: loggedInUserId, role } = req.session.user;
+
     try {
+      const investment = await Investment.findOne({
+        where: { id: investmentId },
+      });
+
+      // CHECK ROLEE SAMA USER
+      if (role !== "admin" && investment.UserId !== loggedInUserId) {
+        let error = `no access!`;
+        return res.redirect(`/companies?error=${error}`);
+      }
+      //
       await Investment.destroy({ where: { id: investmentId } });
 
       res.redirect("/investments");
